# server/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: readworld app example
    description: readworld app example
    version: 0.0.1
host: localhost:8080
basePath: api
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /api/user/update:
        post:
            tags:
                - user
            summary: updateUser user
            operationId: user#updateUser
            parameters:
                - name: UpdateUserRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/UserUpdateUserRequestBody'
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/UserUpdateUserResponseBody'
                        required:
                            - user
            schemes:
                - http
definitions:
    User:
        title: User
        type: object
        properties:
            bio:
                type: string
                example: Dolore praesentium dicta vel facilis reprehenderit.
            email:
                type: string
                example: Ea laborum qui laudantium vero voluptas molestiae.
            image:
                type: string
                example: Sed quasi.
            token:
                type: string
                example: Quae voluptatem quasi dignissimos quod exercitationem.
            username:
                type: string
                example: Animi dolorem aut perferendis nesciunt aliquid.
        example:
            bio: Quia neque.
            email: Ullam ut ex minus.
            image: Quos omnis magni iste explicabo.
            token: Nam ipsam reiciendis sit autem.
            username: Nisi consequatur quo.
        required:
            - email
            - token
            - username
            - bio
            - image
...
    UserUpdateUserRequestBody:
        title: UserUpdateUserRequestBody
        type: object
        properties:
            bio:
                type: string
                example: krr
                maxLength: 4096
            email:
                type: string
                example: cedrick@skiles.name
                format: email
            image:
                type: string
                example: http://ym
                pattern: ^http://.+$
            password:
                type: string
                example: bds
                minLength: 6
                maxLength: 128
            username:
                type: string
                example: 0ZN
                pattern: ^[a-zA-Z0-9_]{3,32}$
        example:
            bio: f09
            email: tre.nicolas@deckowfisher.name
            image: http://a
            password: spo
            username: _P3z
    UserUpdateUserResponseBody:
        title: UserUpdateUserResponseBody
        type: object
        properties:
            user:
                $ref: '#/definitions/User'
        example:
            user:
                bio: Placeat perspiciatis molestiae.
                email: Aut aliquid corrupti quae.
                image: Ut ipsum.
                token: Voluptates consequatur ab sint sit repellendus.
                username: Officia ut necessitatibus aut et maiores.
        required:
            - user